refactor: plugins migration * project * toggleterm
