+++ lua/john/plugins/init.lua
@@ -1,8 +1,6 @@
 require "john.plugins.cmp"
 require "john.plugins.telescope"
 require "john.plugins.nvim-tree"
-require "john.plugins.toggleterm"
-require "john.plugins.project"
 require "john.plugins.indentline"
 require "john.plugins.whichkey"
 require "john.plugins.dap_config"
